requirements.txt
streamlit
plotly
streamlit-plotly-events
google-cloud-bigquery
requests
pandas
python-dotenv
dbt-bigquery
